refactor: move SidebarToggle to base folder
- Move SidebarToggle.tsx to base folder since it's a base component
- Update import paths in BaseSidebar and test files
- All base components now properly organized in base folder
- SidebarToggle will be reused by all sidebar types (story, grammar, etc.)
